Resolve DEBT-230: Decompose seed.ts into focused modules
