+++ scripts/seed.ts
@@ -1,233 +1,17 @@
-import { readFile } from 'node:fs/promises';
 import path from 'node:path';
+import { pathToFileURL } from 'node:url';
 import dotenv from 'dotenv';
-import { and, eq, inArray, like } from 'drizzle-orm';
-import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import { drizzle } from 'drizzle-orm/postgres-js';
-import fg from 'fast-glob';
-import matter from 'gray-matter';
 import postgres from 'postgres';
 import * as schema from '../db/schema';
-import {
-  CANONICAL_SUBSTANCE_SLUGS,
-  CANONICAL_TOPIC_SLUGS,
-  CANONICAL_TREATMENT_SLUGS,
-} from '../lib/content/draftTaxonomy';
-import {
-  canonicalizeMarkdown,
-  canonicalJsonString,
-  parseMdxQuestionBody,
-  sha256Hex,
-} from '../lib/content/parseMdxQuestion';
-import {
-  FullQuestionSchema,
-  QuestionFrontmatterSchema,
-} from '../lib/content/schemas';
-import { computeChoiceSyncPlan, parseChoiceExplanations } from './seed-helpers';
-
-type SeedTag = {
-  slug: string;
-  name: string;
-  kind: schema.TagKind;
-};
+import { readSeedQuestionFiles } from './seed/file-reader';
+import { archivePlaceholderQuestions } from './seed/placeholder-archiver';
+import { syncQuestionsFromFiles } from './seed/question-syncer';
 
 dotenv.config({ path: '.env.local' });
 dotenv.config({ path: '.env' });
 
-type SeedChoice = {
-  label: string;
-  text_md: string;
-  is_correct: boolean;
-  explanation_md: string | null;
-  sort_order: number;
-};
-
-type SeedQuestionRep = {
-  slug: string;
-  stem_md: string;
-  explanation_md: string;
-  difficulty: schema.QuestionDifficulty;
-  status: schema.QuestionStatus;
-  choices: SeedChoice[];
-  tags: SeedTag[];
-};
-
-const CANONICAL_TOPIC_SLUG_SET = new Set<string>(CANONICAL_TOPIC_SLUGS);
-const CANONICAL_SUBSTANCE_SLUG_SET = new Set<string>(CANONICAL_SUBSTANCE_SLUGS);
-const CANONICAL_TREATMENT_SLUG_SET = new Set<string>(CANONICAL_TREATMENT_SLUGS);
-
-function buildSeedRepFromFile(full: unknown): SeedQuestionRep {
-  const parsed = FullQuestionSchema.parse(full);
-  const parsedExplanations = parseChoiceExplanations(parsed.explanationMd);
-  const generalExplanation = parsedExplanations.generalExplanation;
-
-  const sortedTags = [...parsed.frontmatter.tags].sort((a, b) =>
-    a.slug.localeCompare(b.slug),
-  );
-  const sortedChoices = [...parsed.frontmatter.choices].sort((a, b) =>
-    a.label.localeCompare(b.label),
-  );
-  const validLabels = new Set(sortedChoices.map((choice) => choice.label));
-  for (const label of parsedExplanations.perChoice.keys()) {
-    if (!validLabels.has(label)) {
-      throw new Error(
-        `Explanation references choice label "${label}" that is not present in choices for slug "${parsed.frontmatter.slug}"`,
-      );
-    }
-  }
-
-  return {
-    slug: parsed.frontmatter.slug,
-    stem_md: canonicalizeMarkdown(parsed.stemMd),
-    explanation_md: generalExplanation,
-    difficulty: parsed.frontmatter.difficulty,
-    status: parsed.frontmatter.status,
-    tags: sortedTags.map((t) => ({
-      slug: t.slug,
-      name: t.name,
-      kind: t.kind,
-    })),
-    choices: sortedChoices.map((c, index) => ({
-      label: c.label,
-      text_md: canonicalizeMarkdown(c.text),
-      is_correct: c.correct,
-      explanation_md: parsedExplanations.perChoice.get(c.label) ?? null,
-      sort_order: index + 1,
-    })),
-  };
-}
-
-function buildSeedRepFromDb(
-  question: schema.Question,
-  choices: schema.Choice[],
-  tags: SeedTag[],
-): SeedQuestionRep {
-  return {
-    slug: question.slug,
-    stem_md: canonicalizeMarkdown(question.stemMd),
-    explanation_md: canonicalizeMarkdown(question.explanationMd),
-    difficulty: question.difficulty,
-    status: question.status,
-    choices: [...choices]
-      .sort((a, b) => a.sortOrder - b.sortOrder)
-      .map((c) => ({
-        label: c.label,
-        text_md: canonicalizeMarkdown(c.textMd),
-        is_correct: c.isCorrect,
-        explanation_md: c.explanationMd
-          ? canonicalizeMarkdown(c.explanationMd)
-          : null,
-        sort_order: c.sortOrder,
-      })),
-    tags: [...tags].sort((a, b) => a.slug.localeCompare(b.slug)),
-  };
-}
-
-async function upsertTags(
-  tx: PostgresJsDatabase<typeof schema>,
-  incomingTags: SeedTag[],
-): Promise<Map<string, { id: string } & SeedTag>> {
-  const tagSlugs = incomingTags.map((t) => t.slug);
-
-  const existing = tagSlugs.length
-    ? await tx
-        .select()
-        .from(schema.tags)
-        .where(inArray(schema.tags.slug, tagSlugs))
-    : [];
-
-  const bySlug = new Map(existing.map((t) => [t.slug, t]));
-
-  for (const tag of incomingTags) {
-    const found = bySlug.get(tag.slug);
-    if (found) {
-      if (found.name !== tag.name || found.kind !== tag.kind) {
-        throw new Error(
-          `Tag slug "${tag.slug}" already exists but differs (expected name="${tag.name}", kind="${tag.kind}"; got name="${found.name}", kind="${found.kind}")`,
-        );
-      }
-      continue;
-    }
-
-    const [inserted] = await tx
-      .insert(schema.tags)
-      .values({
-        slug: tag.slug,
-        name: tag.name,
-        kind: tag.kind,
-      })
-      .returning();
-
-    if (!inserted) {
-      throw new Error(`Failed to insert tag slug "${tag.slug}"`);
-    }
-
-    bySlug.set(inserted.slug, inserted);
-  }
-
-  return bySlug;
-}
-
-function parseFrontmatterOrThrow(data: unknown) {
-  return QuestionFrontmatterSchema.parse(data);
-}
-
-export function validateSeedQuestionTags(input: {
-  slug: string;
-  tags: Array<Omit<SeedTag, 'kind'> & { kind: SeedTag['kind'] | 'domain' }>;
-}): void {
-  for (const tag of input.tags) {
-    if (tag.kind === 'domain') {
-      throw new Error(
-        `Question "${input.slug}" has domain tag "${tag.slug}" which is not allowed`,
-      );
-    }
-  }
-
-  const topicCount = input.tags.filter((tag) => tag.kind === 'topic').length;
-  const substanceCount = input.tags.filter(
-    (tag) => tag.kind === 'substance',
-  ).length;
-
-  if (topicCount < 1) {
-    throw new Error(
-      `Question "${input.slug}" must have at least one topic tag`,
-    );
-  }
-  if (substanceCount < 1) {
-    throw new Error(
-      `Question "${input.slug}" must have at least one substance tag`,
-    );
-  }
-
-  for (const tag of input.tags) {
-    if (tag.kind === 'topic' && !CANONICAL_TOPIC_SLUG_SET.has(tag.slug)) {
-      throw new Error(
-        `Question "${input.slug}" has non-canonical topic slug "${tag.slug}"`,
-      );
-    }
-
-    if (
-      tag.kind === 'substance' &&
-      !CANONICAL_SUBSTANCE_SLUG_SET.has(tag.slug)
-    ) {
-      throw new Error(
-        `Question "${input.slug}" has non-canonical substance slug "${tag.slug}"`,
-      );
-    }
-
-    if (
-      tag.kind === 'treatment' &&
-      !CANONICAL_TREATMENT_SLUG_SET.has(tag.slug)
-    ) {
-      throw new Error(
-        `Question "${input.slug}" has non-canonical treatment slug "${tag.slug}"`,
-      );
-    }
-  }
-}
-
-async function main(): Promise<void> {
+export async function runSeed(): Promise<void> {
   const databaseUrl = process.env.DATABASE_URL;
   if (!databaseUrl) {
     throw new Error(
@@ -235,244 +19,23 @@ async function main(): Promise<void> {
     );
   }
 
+  const includePlaceholders = process.env.SEED_INCLUDE_PLACEHOLDERS === 'true';
   const sql = postgres(databaseUrl, { max: 1 });
   const db = drizzle(sql, { schema });
+
   try {
-    const includePlaceholders =
-      process.env.SEED_INCLUDE_PLACEHOLDERS === 'true';
-    const patterns = includePlaceholders
-      ? ['content/questions/**/*.mdx']
-      : [
-          'content/questions/**/*.mdx',
-          '!content/questions/placeholder/**/*.mdx',
-        ];
-
-    const files = await fg(patterns, {
-      onlyFiles: true,
-      unique: true,
-      absolute: true,
-      dot: false,
-    });
-
-    if (files.length === 0) {
-      throw new Error(
-        includePlaceholders
-          ? 'No question files found at content/questions/**/*.mdx. Seed requires at least one MDX file.'
-          : 'No question files found after excluding placeholders. Re-run with SEED_INCLUDE_PLACEHOLDERS=true or generate imported content under content/questions/imported/.',
-      );
-    }
-
-    let inserted = 0;
-    let updated = 0;
-    let skipped = 0;
-
-    for (const file of files) {
-      const raw = await readFile(file, 'utf8');
-      const { data, content } = matter(raw);
-
-      const frontmatter = parseFrontmatterOrThrow(data);
-      const { stemMd, explanationMd } = parseMdxQuestionBody(content);
-
-      const seedFromFile = buildSeedRepFromFile({
-        frontmatter,
-        stemMd,
-        explanationMd,
-      });
-      validateSeedQuestionTags({
-        slug: seedFromFile.slug,
-        tags: seedFromFile.tags,
-      });
-
-      const fileHash = sha256Hex(canonicalJsonString(seedFromFile));
-
-      const existing = await db
-        .select()
-        .from(schema.questions)
-        .where(eq(schema.questions.slug, seedFromFile.slug))
-        .limit(1);
-      const existingQuestion = existing.at(0);
-
-      if (!existingQuestion) {
-        await db.transaction(async (tx) => {
-          const [createdQuestion] = await tx
-            .insert(schema.questions)
-            .values({
-              slug: seedFromFile.slug,
-              stemMd: seedFromFile.stem_md,
-              explanationMd: seedFromFile.explanation_md,
-              difficulty: seedFromFile.difficulty,
-              status: seedFromFile.status,
-            })
-            .returning({ id: schema.questions.id });
-
-          if (!createdQuestion) {
-            throw new Error(
-              `Failed to insert question for slug "${seedFromFile.slug}"`,
-            );
-          }
-
-          await tx.insert(schema.choices).values(
-            seedFromFile.choices.map((c) => ({
-              questionId: createdQuestion.id,
-              label: c.label,
-              textMd: c.text_md,
-              isCorrect: c.is_correct,
-              explanationMd: c.explanation_md,
-              sortOrder: c.sort_order,
-            })),
-          );
-
-          const tagMap = await upsertTags(tx, seedFromFile.tags);
-          await tx.insert(schema.questionTags).values(
-            seedFromFile.tags.map((t) => ({
-              questionId: createdQuestion.id,
-              tagId:
-                tagMap.get(t.slug)?.id ??
-                (() => {
-                  throw new Error(`Missing tag id for slug "${t.slug}"`);
-                })(),
-            })),
-          );
-        });
-
-        inserted += 1;
-        continue;
-      }
-
-      const existingChoices = await db
-        .select()
-        .from(schema.choices)
-        .where(eq(schema.choices.questionId, existingQuestion.id));
-
-      const existingTags = await db
-        .select({
-          slug: schema.tags.slug,
-          name: schema.tags.name,
-          kind: schema.tags.kind,
-        })
-        .from(schema.questionTags)
-        .innerJoin(schema.tags, eq(schema.questionTags.tagId, schema.tags.id))
-        .where(eq(schema.questionTags.questionId, existingQuestion.id));
-
-      const seedFromDb = buildSeedRepFromDb(
-        existingQuestion,
-        existingChoices,
-        existingTags,
-      );
-
-      const dbHash = sha256Hex(canonicalJsonString(seedFromDb));
-      if (dbHash === fileHash) {
-        skipped += 1;
-        continue;
-      }
-
-      const desiredLabels = new Set(seedFromFile.choices.map((c) => c.label));
-      const deleteCandidates = existingChoices.filter(
-        (c) => !desiredLabels.has(c.label),
-      );
-
-      const referencedChoiceIds = new Set<string>();
-      if (deleteCandidates.length > 0) {
-        const deleteCandidateIds = deleteCandidates.map((c) => c.id);
-
-        const referenced = await db
-          .select({ selectedChoiceId: schema.attempts.selectedChoiceId })
-          .from(schema.attempts)
-          .where(
-            and(
-              eq(schema.attempts.questionId, existingQuestion.id),
-              inArray(schema.attempts.selectedChoiceId, deleteCandidateIds),
-            ),
-          );
-
-        for (const row of referenced) {
-          if (row.selectedChoiceId)
-            referencedChoiceIds.add(row.selectedChoiceId);
-        }
-      }
-
-      const { deleteChoiceIds } = computeChoiceSyncPlan({
-        existingChoices: existingChoices.map((c) => ({
-          id: c.id,
-          label: c.label,
-        })),
-        desiredChoices: seedFromFile.choices.map((c) => ({ label: c.label })),
-        referencedChoiceIds,
-      });
-
-      await db.transaction(async (tx) => {
-        await tx
-          .update(schema.questions)
-          .set({
-            stemMd: seedFromFile.stem_md,
-            explanationMd: seedFromFile.explanation_md,
-            difficulty: seedFromFile.difficulty,
-            status: seedFromFile.status,
-            updatedAt: new Date(),
-          })
-          .where(eq(schema.questions.id, existingQuestion.id));
-
-        if (deleteChoiceIds.length > 0) {
-          await tx
-            .delete(schema.choices)
-            .where(inArray(schema.choices.id, deleteChoiceIds));
-        }
-
-        for (const choice of seedFromFile.choices) {
-          await tx
-            .insert(schema.choices)
-            .values({
-              questionId: existingQuestion.id,
-              label: choice.label,
-              textMd: choice.text_md,
-              isCorrect: choice.is_correct,
-              explanationMd: choice.explanation_md,
-              sortOrder: choice.sort_order,
-            })
-            .onConflictDoUpdate({
-              target: [schema.choices.questionId, schema.choices.label],
-              set: {
-                textMd: choice.text_md,
-                isCorrect: choice.is_correct,
-                explanationMd: choice.explanation_md,
-                sortOrder: choice.sort_order,
-              },
-            });
-        }
-
-        await tx
-          .delete(schema.questionTags)
-          .where(eq(schema.questionTags.questionId, existingQuestion.id));
-
-        const tagMap = await upsertTags(tx, seedFromFile.tags);
-        await tx.insert(schema.questionTags).values(
-          seedFromFile.tags.map((t) => ({
-            questionId: existingQuestion.id,
-            tagId:
-              tagMap.get(t.slug)?.id ??
-              (() => {
-                throw new Error(`Missing tag id for slug "${t.slug}"`);
-              })(),
-          })),
-        );
-      });
-
-      updated += 1;
-    }
+    const files = await readSeedQuestionFiles(includePlaceholders);
+    const counts = await syncQuestionsFromFiles(db, files);
 
     console.info(
-      `Seed complete: inserted=${inserted} updated=${updated} skipped=${skipped} (files=${files.length})`,
+      `Seed complete: inserted=${counts.inserted} updated=${counts.updated} skipped=${counts.skipped} (files=${files.length})`,
     );
     console.info(`Content root: ${path.resolve('content/questions')}`);
 
     if (!includePlaceholders) {
-      const archived = await db
-        .update(schema.questions)
-        .set({ status: 'archived', updatedAt: new Date() })
-        .where(like(schema.questions.slug, 'placeholder-%'))
-        .returning({ id: schema.questions.id });
+      const archivedCount = await archivePlaceholderQuestions(db);
       console.info(
-        `Archived placeholders: ${archived.length} (slug LIKE "placeholder-%")`,
+        `Archived placeholders: ${archivedCount} (slug LIKE "placeholder-%")`,
       );
     }
   } finally {
@@ -480,7 +43,16 @@ async function main(): Promise<void> {
   }
 }
 
-main().catch((error) => {
-  console.error(error);
-  process.exitCode = 1;
-});
+async function main(): Promise<void> {
+  await runSeed();
+}
+
+if (
+  process.argv[1] &&
+  import.meta.url === pathToFileURL(process.argv[1]).href
+) {
+  main().catch((error) => {
+    console.error(error);
+    process.exitCode = 1;
+  });
+}
